fix: fix bug in searcher

Created-by: huohuixin
Commit-by: huohuixin
Merged-by: CangjiePL
Description: ## 变更内容（必填）

fix: 修复 fileid转换逻辑，确保正确处理 超过INT_MAX的fileid

## 变更类型（必填）

**请在保存后点击复选框或在编辑时将对应项前面的 `[ ]` 改为 `[x]` 。**

- [ ] 新增需求
- [x] 问题修复
- [ ] 构建过程或辅助工具变动
- [ ] 文档更新

## 变更内容自检（必填）

**请勿修改或删除以下选项内容，请在保存后点击复选框或在编辑时将对应项前面的 `[ ]` 改为 `[x]` 。**

### 编译本地自验证结果：

- [x] 仓颉编译器编译通过
- [x] 仓颉运行时编译通过
- [x] 仓颉标准库编译通过
- [ ] 不涉及，辅助工具变动或者文档更新可以直接选择此项

### 测试用例本地自验证结果：

- [x] 通过，请在下方提供截图证明
修复前ut用例
![image.png](https://raw.gitcode.com/user-images/assets/7239667/aff6f5c8-8195-4a79-8730-8a6598779e2b/image.png 'image.png')
修复后ut
![image.png](https://raw.gitcode.com/user-images/assets/7239667/89f86fff-7d31-4dd0-bea2-ec3796c2673b/image.png 'image.png')
- [ ] 不涉及，请在下方说明情况

## 关联的issue（必填）

https://gitcode.com/Cangjie/cangjie_tools/issues/341

See merge request: Cangjie/cangjie_compiler!1081

+++ src/AST/Searcher.cpp
@@ -161,7 +161,7 @@ uint32_t PosSearchApi::MAX_DIGITS_COLUMN = 5; /**< Max column width is 99999. */
 
 void PosSearchApi::UpdatePosLimit(unsigned int fileId, int line, int column)
 {
-    uint32_t digitsFile = static_cast<uint32_t>(FillZero(static_cast<int>(fileId), 0).size());
+    uint32_t digitsFile = static_cast<uint32_t>(FillZero(fileId, 0).size());
     uint32_t digitsLine = static_cast<uint32_t>(FillZero(line, 0).size());
     uint32_t digitsColumn = static_cast<uint32_t>(FillZero(column, 0).size());
     if (digitsFile > MAX_DIGITS_FILE) {
@@ -177,7 +177,7 @@ void PosSearchApi::UpdatePosLimit(unsigned int fileId, int line, int column)
 
 std::string PosSearchApi::PosToStr(const Position& pos)
 {
-    std::string ret = FillZero(static_cast<int>(pos.fileID), static_cast<int>(MAX_DIGITS_FILE));
+    std::string ret = FillZero(pos.fileID, static_cast<int>(MAX_DIGITS_FILE));
     ret += FillZero(pos.line, static_cast<int>(MAX_DIGITS_LINE));
     ret += FillZero(pos.column, static_cast<int>(MAX_DIGITS_COLUMN));
     return ret;
